scope and namespace cleanup.
drupalBookmarklet function is a constructor that takes Drupal host and path
arguments.

all methods are prototypes.

use jQuery proxy to give handleMessage method context of bookmarklet instance.

bookmarklet itself creates a global variable that is a hash of the Drupal site's
base URL. This should allow users to invoke the script for multiple Drupal sites
without refreshing the page.